+++ ssbh_data_py_derive/src/lib.rs
@@ -166,7 +166,7 @@ pub fn map_py_derive(input: TokenStream) -> TokenStream {
             let named_fields: Vec<_> = fields.named.iter().map(|field| &field.ident).collect();
             quote! {
                 #(
-                    #named_fields: self.#named_fields.map_py(py, use_numpy)?
+                    #named_fields: self.#named_fields.map_py(py)?
                 ),*
             }
         }
@@ -183,7 +183,6 @@ fn generate_map_py(name: &Ident, map_type: &syn::Path, map_data: &TokenStream2) 
             fn map_py(
                 &self,
                 py: pyo3::Python,
-                use_numpy: bool
             ) -> pyo3::prelude::PyResult<#map_type> {
                 Ok(
                     #map_type {
@@ -198,7 +197,6 @@ fn generate_map_py(name: &Ident, map_type: &syn::Path, map_data: &TokenStream2) 
             fn map_py(
                 &self,
                 py: pyo3::Python,
-                use_numpy: bool
             ) -> pyo3::prelude::PyResult<#name> {
                 Ok(
                     #name {
@@ -213,17 +211,17 @@ fn generate_map_py(name: &Ident, map_type: &syn::Path, map_data: &TokenStream2) 
             fn map_py(
                 &self,
                 py: pyo3::Python,
-                use_numpy: bool
+
             ) -> pyo3::prelude::PyResult<pyo3::PyObject> {
-                let x: #name = self.map_py(py, use_numpy)?;
+                let x: #name = self.map_py(py)?;
                 Ok(x.into_py(py))
             }
         }
 
         impl crate::MapPy<#map_type> for pyo3::PyObject {
-            fn map_py(&self, py: pyo3::Python, use_numpy: bool) -> pyo3::prelude::PyResult<#map_type> {
+            fn map_py(&self, py: pyo3::Python, ) -> pyo3::prelude::PyResult<#map_type> {
                 let x: #name = self.extract(py)?;
-                x.map_py(py, use_numpy)
+                x.map_py(py)
             }
         }
     }