+++ ssbh_data_py_types/src/mesh_data.rs
@@ -1,6 +1,7 @@
 use crate::create_py_list_from_slice;
 use crate::{MapPy, PyInit, PyRepr, Pyi, PyiMethods};
 use num_traits::AsPrimitive;
+use numpy::PyArrayMethods;
 use numpy::{ndarray::Dim, IntoPyArray, PyArray, PyArray2};
 use pyo3::exceptions::PyValueError;
 use pyo3::{create_exception, wrap_pyfunction};
@@ -57,7 +58,7 @@ impl MeshData {
     }
 
     fn save(&self, py: Python, path: &str) -> PyResult<()> {
-        self.map_py(py, false)?
+        self.map_py(py)?
             .write_to_file(path)
             .map_err(|e| MeshDataError::new_err(format!("{}", e)))
     }
@@ -113,7 +114,7 @@ pub struct MeshObjectData {
 
     #[pyo3(get, set)]
     #[pyinit(default = "PyList::empty(py).into()")]
-    #[pyi(default = "[]", python_type = "list[int]")]
+    #[pyi(default = "numpy.array([])", python_type = "numpy.ndarray")]
     pub vertex_indices: PyObject,
 
     #[pyo3(get, set)]
@@ -187,90 +188,40 @@ pub struct AttributeData {
 
     #[pyo3(get, set)]
     #[pyinit(default = "PyList::empty(py).into()")]
-    #[pyi(default = "[]", python_type = "list[list[float]]")]
+    #[pyi(default = "numpy.array([])", python_type = "numpy.ndarray")]
     pub data: PyObject,
 }
 
 impl MapPy<PyObject> for VectorDataRs {
-    fn map_py(&self, py: Python, use_numpy: bool) -> PyResult<PyObject> {
-        if !use_numpy {
-            match self {
-                VectorDataRs::Vector2(v) => create_py_list_from_slice(py, v).map(Into::into),
-                VectorDataRs::Vector3(v) => create_py_list_from_slice(py, v).map(Into::into),
-                VectorDataRs::Vector4(v) => create_py_list_from_slice(py, v).map(Into::into),
-            }
-        } else {
-            // This gives a roughly 4.7x speedup for reading a 25 MB mesh (gamewatch).
-            // TODO: Can we avoid flattening and then reshaping?
-            // TODO: Handle errors?
-            // Ok(match self {
-            //     VectorDataRs::Vector2(v) => numpy::ndarray::ArrayBase::from_shape_vec(
-            //         (v.len(), 2),
-            //         v.iter().copied().flatten().collect::<Vec<f32>>(),
-            //     )
-            //     .unwrap()
-            //     .into_pyarray(py)
-            //     .into(),
-            //     VectorDataRs::Vector3(v) => numpy::ndarray::ArrayBase::from_shape_vec(
-            //         (v.len(), 3),
-            //         v.iter().copied().flatten().collect::<Vec<f32>>(),
-            //     )
-            //     .unwrap()
-            //     .into_pyarray(py)
-            //     .into(),
-            //     VectorDataRs::Vector4(v) => numpy::ndarray::ArrayBase::from_shape_vec(
-            //         (v.len(), 4),
-            //         v.iter().copied().flatten().collect::<Vec<f32>>(),
-            //     )
-            //     .unwrap()
-            //     .into_pyarray(py)
-            //     .into(),
-            // })
-            todo!()
+    fn map_py(&self, py: Python) -> PyResult<PyObject> {
+        match self {
+            VectorDataRs::Vector2(v) => vectors_pyarray(py, v),
+            VectorDataRs::Vector3(v) => vectors_pyarray(py, v),
+            VectorDataRs::Vector4(v) => vectors_pyarray(py, v),
         }
     }
 }
 
-fn vector_data<T: AsPrimitive<f32> + numpy::Element>(
-    arr: &PyArray<T, Dim<[usize; 2]>>,
-) -> PyResult<VectorDataRs> {
-    // Use AsPrimitive to allow truncating types like f64.
-    // TODO: Avoid unwrap?
-    todo!()
-    // match arr.dims()[1] {
-    //     2 => Ok(VectorDataRs::Vector2(
-    //         arr.readonly()
-    //             .as_slice()
-    //             .unwrap()
-    //             .chunks_exact(2)
-    //             .map(|v| [v[0].as_(), v[1].as_()])
-    //             .collect(),
-    //     )),
-    //     3 => Ok(VectorDataRs::Vector3(
-    //         arr.readonly()
-    //             .as_slice()
-    //             .unwrap()
-    //             .chunks_exact(3)
-    //             .map(|v| [v[0].as_(), v[1].as_(), v[2].as_()])
-    //             .collect(),
-    //     )),
-    //     4 => Ok(VectorDataRs::Vector4(
-    //         arr.readonly()
-    //             .as_slice()
-    //             .unwrap()
-    //             .chunks_exact(4)
-    //             .map(|v| [v[0].as_(), v[1].as_(), v[2].as_(), v[3].as_()])
-    //             .collect(),
-    //     )),
-    //     _ => Err(PyValueError::new_err(format!(
-    //         "Cannot create VectorData from an array of shape {:?}.",
-    //         arr.shape()
-    //     ))),
-    // }
+fn vectors_pyarray<const N: usize>(py: Python, values: &[[f32; N]]) -> PyResult<PyObject> {
+    // This flatten will be optimized in Release mode.
+    // This avoids needing unsafe code.
+    // TODO: Can we avoid flattening and then reshaping?
+    // TODO: Handle errors?
+    let count = values.len();
+    Ok(values
+        .iter()
+        .flat_map(|v| v)
+        .copied()
+        .collect::<Vec<f32>>()
+        .into_pyarray(py)
+        .reshape((count, N))
+        .unwrap()
+        .into_any()
+        .into())
 }
 
 impl MapPy<VectorDataRs> for PyObject {
-    fn map_py(&self, py: Python, _use_numpy: bool) -> PyResult<VectorDataRs> {
+    fn map_py(&self, py: Python) -> PyResult<VectorDataRs> {
         // We don't know the type from Python at this point.
         // Try all the supported types and fail if all conversions fail.
         // TODO: Is there an easy way to convert f64 to f32 for the entire array?
@@ -287,26 +238,26 @@ impl MapPy<VectorDataRs> for PyObject {
 }
 
 #[pyfunction]
-fn read_mesh(py: Python, path: &str, use_numpy: Option<bool>) -> PyResult<MeshData> {
+fn read_mesh(py: Python, path: &str) -> PyResult<MeshData> {
     ssbh_data::mesh_data::MeshData::from_file(path)
         .map_err(|e| MeshDataError::new_err(format!("{}", e)))?
-        .map_py(py, use_numpy.unwrap_or(false))
+        .map_py(py)
 }
 
 #[pyfunction]
 fn transform_points(py: Python, points: PyObject, transform: PyObject) -> PyResult<PyObject> {
-    let points = points.map_py(py, false)?;
-    let transform = transform.map_py(py, false)?;
+    let points = points.map_py(py)?;
+    let transform = transform.map_py(py)?;
     let transformed_points = ssbh_data::mesh_data::transform_points(&points, &transform);
-    transformed_points.map_py(py, false)
+    transformed_points.map_py(py)
 }
 
 #[pyfunction]
 fn transform_vectors(py: Python, points: PyObject, transform: PyObject) -> PyResult<PyObject> {
-    let points = points.map_py(py, false)?;
-    let transform = transform.map_py(py, false)?;
+    let points = points.map_py(py)?;
+    let transform = transform.map_py(py)?;
     let transformed_points = ssbh_data::mesh_data::transform_vectors(&points, &transform);
-    transformed_points.map_py(py, false)
+    transformed_points.map_py(py)
 }
 
 #[pyfunction]
@@ -315,7 +266,7 @@ fn calculate_smooth_normals(
     positions: PyObject,
     vertex_indices: PyObject,
 ) -> PyResult<Py<PyList>> {
-    let positions = positions.map_py(py, false)?;
+    let positions = positions.map_py(py)?;
     let vertex_indices = vertex_indices.extract::<Vec<u32>>(py)?;
     let normals = ssbh_data::mesh_data::calculate_smooth_normals(&positions, &vertex_indices);
     create_py_list_from_slice(py, &normals)
@@ -329,9 +280,9 @@ fn calculate_tangents_vec4(
     uvs: PyObject,
     vertex_indices: PyObject,
 ) -> PyResult<Py<PyList>> {
-    let positions = positions.map_py(py, false)?;
-    let normals = normals.map_py(py, false)?;
-    let uvs = uvs.map_py(py, false)?;
+    let positions = positions.map_py(py)?;
+    let normals = normals.map_py(py)?;
+    let uvs = uvs.map_py(py)?;
 
     let vertex_indices = vertex_indices.extract::<Vec<u32>>(py)?;
     let tangents =