+++ ssbh_data_py_types/src/modl_data.rs
@@ -65,9 +65,7 @@ impl ModlData {
     }
 
     fn save(&self, py: Python, path: &str) -> PyResult<()> {
-        self.map_py(py, false)?
-            .write_to_file(path)
-            .map_err(PyErr::from)
+        self.map_py(py)?.write_to_file(path).map_err(PyErr::from)
     }
 
     fn __repr__(&self) -> String {
@@ -109,5 +107,5 @@ pub struct ModlEntryData {
 fn read_modl(py: Python, path: &str) -> PyResult<ModlData> {
     ssbh_data::modl_data::ModlData::from_file(path)
         .map_err(|e| ModlDataError::new_err(format!("{}", e)))?
-        .map_py(py, false)
+        .map_py(py)
 }