+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## unreleased
+### Changed
+* Enabled numpy arrays by default for better performance and removed `use_numpy` argument from all relevant methods.
+
 ## 0.8.4 - 2024-08-01
 ### Fixed
 - Fixed an issue where `ssbh_data_py.mesh_ex_data.EntryFlags` could not be instantiated from Python.